fix: break plp page
