+++ ui/src/container/index.js
@@ -2,3 +2,5 @@ export { default as Header } from "./Header";
 export { default as Form } from "./Form";
 export { default as Carousel } from "./Carousel";
 export { default as HomePageCategories } from "./HomePageCategories";
+export { default as ProductCategories } from "./ProductCategories";
+export { default as Products } from "./Products";
