--- ui/src/pages/ProductListing/abc.scss
@@ -1,101 +1,0 @@
-@use "../../variables";
-
-.product-listing-page {
-  display: flex;
-  justify-content: space-between;
-  gap: 2%;
-  .categories-list {
-    background-color: variables.$app-grey-color;
-    width: 20%;
-    li {
-      list-style: none;
-      padding: 5% 10%;
-      border-bottom: 1px solid variables.$app-dark-grey-color;
-    }
-  }
-
-  .products {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-    justify-content: space-between;
-    column-gap: 2%;
-    row-gap: 1%;
-    width: 80%;
-    .product {
-      width: 100%;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      margin-top: 5%;
-      margin-right: 2%;
-      padding: 2%;
-      box-shadow: 0 12px 24px 0 rgba(0, 0, 0, 0.2);
-      h3 {
-        margin-top: 1%;
-        color: variables.$color-black;
-        font-weight: bold;
-      }
-      img {
-        width: 100%;
-      }
-      .description {
-        background-color: variables.$app-grey-color;
-        color: variables.$color-black;
-        padding: 2%;
-      }
-      .price {
-        margin: 5% 0 2% 0;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        gap: 10%;
-        p {
-          width: 100%;
-          color: variables.$color-black;
-        }
-      }
-    }
-  }
-}
-
-@media screen and (max-width: 1024px) {
-  .product-listing-page {
-    .products {
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-    }
-  }
-}
-
-@media screen and (max-width: 480px) {
-  .product-listing-page {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    gap: 2%;
-    .categories-list {
-      width: 100%;
-      display: relative;
-      li {
-        display: none;
-        &:first-child {
-          background-color: variables.$app-red-color;
-          color: variables.$color-white;
-          display: block;
-        }
-        &::after {
-          content: "\2193";
-          position: absolute;
-          color: variables.$color-white;
-          right: 5%;
-        }
-      }
-    }
-    .products {
-      display: grid;
-      grid-template-columns: 1fr;
-      row-gap: 0;
-      width: 100%;
-    }
-  }
-}
